change the repo structure
